what_i_noted/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-crispy-forms==1.8.0
django-tinymce==2.8.0
idna==2.8
Pillow==6.2.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.6
virtualenv==16.7.7
